Part / PartDesign: move linear and polar pattern logic out of PartDesign and into Part extensions
