+++ src/Mod/Part/App/AppPart.cpp
@@ -94,6 +94,8 @@
 #include "FeaturePartSpline.h"
 #include "FeatureProjectOnSurface.h"
 #include "FeatureRevolution.h"
+#include "LinearPatternExtension.h"
+#include "PolarPatternExtension.h"
 #include "Geometry.h"
 #include "Geometry2d.h"
 #include "GeometryBoolExtensionPy.h"
@@ -446,6 +448,8 @@ PyMOD_INIT_FUNC(Part)
     Part::AttachExtensionPython ::init();
     Part::PreviewExtension      ::init();
     Part::PrismExtension        ::init();
+    Part::LinearPatternExtension::init();
+    Part::PolarPatternExtension ::init();
 
     Part::Feature               ::init();
     Part::FeatureExt            ::init();
