+++ src/Mod/Part/App/PolarPatternExtension.h
@@ -1,0 +1,77 @@
+// SPDX-License-Identifier: LGPL-2.1-or-later
+
+/******************************************************************************
+ *   Copyright (c) 2012 Jan Rheinländer <jrheinlaender@users.sourceforge.net> *
+ *                                                                            *
+ *   This file is part of the FreeCAD CAx development system.                 *
+ *                                                                            *
+ *   This library is free software; you can redistribute it and/or            *
+ *   modify it under the terms of the GNU Library General Public              *
+ *   License as published by the Free Software Foundation; either             *
+ *   version 2 of the License, or (at your option) any later version.         *
+ *                                                                            *
+ *   This library  is distributed in the hope that it will be useful,         *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of           *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the            *
+ *   GNU Library General Public License for more details.                     *
+ *                                                                            *
+ *   You should have received a copy of the GNU Library General Public        *
+ *   License along with this library; see the file COPYING.LIB. If not,       *
+ *   write to the Free Software Foundation, Inc., 59 Temple Place,            *
+ *   Suite 330, Boston, MA  02111-1307, USA                                   *
+ *                                                                            *
+ ******************************************************************************/
+
+
+#pragma once
+
+#include <App/PropertyUnits.h>
+#include <App/PropertyStandard.h>
+#include <App/PropertyLinks.h>
+#include <App/DocumentObjectExtension.h>
+#include <gp_Ax2.hxx>
+#include <gp_Trsf.hxx>
+
+namespace Part
+{
+
+enum class PolarPatternMode
+{
+    Extent,
+    Spacing
+};
+
+class PartExport PolarPatternExtension: public App::DocumentObjectExtension
+{
+    EXTENSION_PROPERTY_HEADER(Part::PolarPatternExtension);
+
+public:
+    PolarPatternExtension();
+    ~PolarPatternExtension() override = default;
+
+    App::PropertyLinkSub Axis;
+    App::PropertyBool Reversed;
+    App::PropertyEnumeration Mode;
+    App::PropertyAngle Angle;
+    App::PropertyAngle Offset;
+    App::PropertyIntegerConstraint Occurrences;
+    App::PropertyFloatList Spacings;
+    App::PropertyFloatList SpacingPattern;
+
+    // Standard helper to calculate the transformations list
+    const std::list<gp_Trsf> calculateTransformations() const;
+
+    // Virtual resolution of the rotation axis definitions
+    virtual gp_Ax2 getRotation() const;
+
+    void setReadWriteStatusForMode(PolarPatternMode mode);
+    void updateSpacings();
+
+    short extensionMustExecute() override;
+    void extensionOnChanged(const App::Property* prop) override;
+
+    static const App::PropertyIntegerConstraint::Constraints intOccurrences;
+    static const char* ModeEnums[];
+};
+
+}  // namespace Part
