+++ src/Mod/PartDesign/App/FeaturePolarPattern.cpp
@@ -46,172 +46,11 @@ using namespace PartDesign;
 namespace PartDesign
 {
 
-
-PROPERTY_SOURCE(PartDesign::PolarPattern, PartDesign::Transformed)
-
-const App::PropertyIntegerConstraint::Constraints PolarPattern::intOccurrences
-    = {1, std::numeric_limits<int>::max(), 1};
-const App::PropertyAngle::Constraints PolarPattern::floatAngle
-    = {Base::toDegrees<double>(Precision::Angular()), 360.0, 1.0};
-
-const char* PolarPattern::ModeEnums[] = {"Extent", "Spacing", nullptr};
+PROPERTY_SOURCE_WITH_EXTENSIONS(PartDesign::PolarPattern, PartDesign::Transformed)
 
 PolarPattern::PolarPattern()
 {
-    auto initialMode = PolarPatternMode::Extent;
-
-    ADD_PROPERTY_TYPE(
-        Axis,
-        (nullptr),
-        "PolarPattern",
-        App::Prop_None,
-        "The axis of rotation for the pattern. This can be a datum axis, a sketch "
-        "axis, a circular edge, or the normal of a planar face."
-    );
-    ADD_PROPERTY_TYPE(
-        Reversed,
-        (0),
-        "PolarPattern",
-        App::Prop_None,
-        "Reverses the pattern direction from counter-clockwise (default) to clockwise."
-    );
-    ADD_PROPERTY_TYPE(
-        Mode,
-        (long(initialMode)),
-        "PolarPattern",
-        App::Prop_None,
-        "Selects how the pattern is dimensioned.\n'Extent': Uses the total angle to contain all "
-        "instances.\n'Spacing': Uses the angle between consecutive instances."
-    );
-    Mode.setEnums(PolarPattern::ModeEnums);
-    ADD_PROPERTY_TYPE(
-        Angle,
-        (360.0),
-        "PolarPattern",
-        App::Prop_None,
-        "The total angle of the pattern, measured from the first to the last "
-        "instance. This is only used when the Mode is set to 'Extent'."
-    );
-    ADD_PROPERTY_TYPE(
-        Offset,
-        (120.0),
-        "PolarPattern",
-        App::Prop_None,
-        "The angular distance between each instance of the pattern. This is only "
-        "used when the Mode is set to 'Spacing'."
-    );
-    ADD_PROPERTY_TYPE(
-        Spacings,
-        ({-1.0, -1.0, -1.0}),
-        "PolarPattern",
-        App::Prop_None,
-        "A list of custom angular spacings between instances. If a value is -1, the "
-        "global 'Offset' is used for that spacing. The list should have one less "
-        "item than the number of occurrences."
-    );
-    ADD_PROPERTY_TYPE(
-        SpacingPattern,
-        ({}),
-        "PolarPattern",
-        App::Prop_None,
-        "(Experimental and subject to change. To enable "
-        "this in the UI you can add a boolean parameter 'ExperiementalFeature' in "
-        "Preferences/Mod/Part)\nDefines a repeating pattern of spacings for the second direction. "
-        "For example, a list of [10, 20] will create alternating spacings of 10mm and 20mm."
-    );
-    ADD_PROPERTY_TYPE(
-        Occurrences,
-        (3),
-        "PolarPattern",
-        App::Prop_None,
-        "The total number of instances in the pattern, including the original feature."
-    );
-    Occurrences.setConstraints(&intOccurrences);
-
-    setReadWriteStatusForMode(initialMode);
-}
-
-short PolarPattern::mustExecute() const
-{
-    if (Axis.isTouched() || Reversed.isTouched() || Mode.isTouched() ||
-        // Angle and Offset are mutually exclusive, only one could be updated at once
-        Angle.isTouched() || Offset.isTouched() || Spacings.isTouched()
-        || SpacingPattern.isTouched() || Occurrences.isTouched()) {
-        return 1;
-    }
-    return Transformed::mustExecute();
-}
-
-const std::list<gp_Trsf> PolarPattern::getTransformations(const std::vector<App::DocumentObject*>)
-{
-    int occurrences = Occurrences.getValue();
-    if (occurrences < 1) {
-        throw Base::ValueError("At least one occurrence required");
-    }
-
-    if (occurrences == 1) {
-        return {gp_Trsf()};
-    }
-
-    gp_Ax2 axis = getRotation();
-
-    double angle;
-
-    if (Mode.getValue() == (int)PolarPatternMode::Extent) {
-        angle = Angle.getValue();
-
-        if (std::fabs(angle - 360.0) < Precision::Confusion()) {
-            angle /= occurrences;  // Because e.g. two occurrences in 360 degrees need to be 180
-                                   // degrees apart
-        }
-        else {
-            angle /= occurrences - 1;
-        }
-
-        angle = Base::toRadians(angle);
-        if (fabs(angle) < Precision::Angular()) {
-            throw Base::ValueError("Pattern angle can't be null");
-        }
-    }
-
-    // make sure spacings are correct size :
-    updateSpacings();
-
-    const std::vector<double> spacings = Spacings.getValues();
-    const std::vector<double> pattern = SpacingPattern.getValues();
-    bool usePattern = pattern.size() > 1;
-
-    double cumulativeSpacings = 0.0;
-
-    std::list<gp_Trsf> transformations;
-    gp_Trsf trans;
-    transformations.push_back(trans);
-
-
-    // Note: The original feature is already included in the list of transformations!
-    // Therefore we start with occurrence number 1
-    for (int i = 1; i < occurrences; i++) {
-        if (Mode.getValue() == (int)PolarPatternMode::Spacing) {
-            double spacing;
-            if (spacings[i - 1] != -1.0) {
-                spacing = spacings[i - 1];
-            }
-            else if (usePattern) {
-                spacing = pattern[(size_t)fmod(i - 1, pattern.size())];
-            }
-            else {
-                spacing = Offset.getValue();
-            }
-            cumulativeSpacings += Base::toRadians(spacing);
-            trans.SetRotation(axis.Axis(), cumulativeSpacings);
-        }
-        else {
-            trans.SetRotation(axis.Axis(), angle * i);
-        }
-        transformations.push_back(trans);
-    }
-
-    return transformations;
+    Part::PolarPatternExtension::initExtension(this);
 }
 
 gp_Ax2 PolarPattern::getRotation() const
@@ -220,117 +59,31 @@ gp_Ax2 PolarPattern::getRotation() const
     if (!refObject) {
         return gp_Ax2();
     }
-    std::vector<std::string> subStrings = Axis.getSubValues();
-    if (subStrings.empty()) {
-        return gp_Ax2();
-    }
 
-    gp_Pnt axbase;
-    gp_Dir axdir;
-    if (refObject->isDerivedFrom<Part::Part2DObject>()) {
-        Part::Part2DObject* refSketch = static_cast<Part::Part2DObject*>(refObject);
-        Base::Axis axis;
-        if (subStrings[0] == "H_Axis") {
-            axis = refSketch->getAxis(Part::Part2DObject::H_Axis);
-            axis *= refSketch->Placement.getValue();
-        }
-        else if (subStrings[0] == "V_Axis") {
-            axis = refSketch->getAxis(Part::Part2DObject::V_Axis);
-            axis *= refSketch->Placement.getValue();
-        }
-        else if (subStrings[0] == "N_Axis") {
-            axis = refSketch->getAxis(Part::Part2DObject::N_Axis);
-            axis *= refSketch->Placement.getValue();
-        }
-        else if (subStrings[0].compare(0, 4, "Axis") == 0) {
-            int AxId = std::atoi(subStrings[0].substr(4, 4000).c_str());
-            if (AxId >= 0 && AxId < refSketch->getAxisCount()) {
-                axis = refSketch->getAxis(AxId);
-            }
-            axis *= refSketch->Placement.getValue();
-        }
-        else if (subStrings[0].compare(0, 4, "Edge") == 0) {
-            Part::TopoShape refShape = refSketch->Shape.getShape();
-            TopoDS_Shape ref = refShape.getSubShape(subStrings[0].c_str());
-            TopoDS_Edge refEdge = TopoDS::Edge(ref);
-            if (refEdge.IsNull()) {
-                throw Base::ValueError("Failed to extract direction edge");
-            }
-            BRepAdaptor_Curve adapt(refEdge);
-            if (adapt.GetType() != GeomAbs_Line) {
-                throw Base::TypeError("Direction edge must be a straight line");
-            }
-
-            gp_Pnt p = adapt.Line().Location();
-            gp_Dir d = adapt.Line().Direction();
-
-            axis.setBase(Base::Vector3d(p.X(), p.Y(), p.Z()));
-            axis.setDirection(Base::Vector3d(d.X(), d.Y(), d.Z()));
-        }
-
-        axbase = gp_Pnt(axis.getBase().x, axis.getBase().y, axis.getBase().z);
-        axdir = gp_Dir(axis.getDirection().x, axis.getDirection().y, axis.getDirection().z);
-    }
-    else if (refObject->isDerivedFrom<PartDesign::Line>()) {
+    if (refObject->isDerivedFrom<PartDesign::Line>()) {
         PartDesign::Line* line = static_cast<PartDesign::Line*>(refObject);
         Base::Vector3d base = line->getBasePoint();
-        axbase = gp_Pnt(base.x, base.y, base.z);
+        gp_Pnt axbase(base.x, base.y, base.z);
         Base::Vector3d dir = line->getDirection();
-        axdir = gp_Dir(dir.x, dir.y, dir.z);
-    }
-    else if (refObject->isDerivedFrom<App::Line>()) {
-        App::Line* line = static_cast<App::Line*>(refObject);
-        Base::Vector3d base = line->getBasePoint();
-        axbase = gp_Pnt(base.x, base.y, base.z);
-        Base::Vector3d d = line->getDirection();
-        axdir = gp_Dir(d.x, d.y, d.z);
-    }
-    else if (refObject->isDerivedFrom<Part::Feature>()) {
-        if (subStrings[0].empty()) {
-            throw Base::ValueError("No axis reference specified");
+        gp_Dir axdir(dir.x, dir.y, dir.z);
+
+        TopLoc_Location invObjLoc = this->getLocation().Inverted();
+        axbase.Transform(invObjLoc.Transformation());
+        axdir.Transform(invObjLoc.Transformation());
+
+        gp_Ax2 axis(axbase, axdir);
+        if (Reversed.getValue()) {
+            axis.SetDirection(axis.Direction().Reversed());
         }
-        Part::Feature* refFeature = static_cast<Part::Feature*>(refObject);
-        Part::TopoShape refShape = refFeature->Shape.getShape();
-        TopoDS_Shape ref = refShape.getSubShape(subStrings[0].c_str());
-
-        if (ref.ShapeType() == TopAbs_EDGE) {
-            TopoDS_Edge refEdge = TopoDS::Edge(ref);
-            if (refEdge.IsNull()) {
-                throw Base::ValueError("Failed to extract axis edge");
-            }
-            BRepAdaptor_Curve adapt(refEdge);
-            if (adapt.GetType() == GeomAbs_Line) {
-                axbase = adapt.Line().Location();
-                axdir = adapt.Line().Direction();
-            }
-            else if (adapt.GetType() == GeomAbs_Circle) {
-                axbase = adapt.Circle().Location();
-                axdir = adapt.Circle().Axis().Direction();
-            }
-            else {
-                throw Base::TypeError("Rotation edge must be a straight line, circle or arc of circle");
-            }
-        }
-        else {
-            throw Base::TypeError("Axis reference must be an edge");
-        }
-    }
-    else {
-        throw Base::TypeError("Axis reference must be edge of a feature or datum line");
+        return axis;
     }
 
-    // Transform the axis into the local coordinate system of this feature.
-    TopLoc_Location invObjLoc = this->getLocation().Inverted();
-    axbase.Transform(invObjLoc.Transformation());
-    axdir.Transform(invObjLoc.Transformation());
+    return Part::PolarPatternExtension::getRotation();
+}
 
-    gp_Ax2 axis(axbase, axdir);
-
-    if (Reversed.getValue()) {
-        axis.SetDirection(axis.Direction().Reversed());
-    }
-
-    return axis;
+const std::list<gp_Trsf> PolarPattern::getTransformations(const std::vector<App::DocumentObject*>)
+{
+    return calculateTransformations();
 }
 
 void PolarPattern::handleChangedPropertyType(
@@ -351,51 +104,6 @@ void PolarPattern::handleChangedPropertyType(
     else {
         Transformed::handleChangedPropertyType(reader, TypeName, prop);
     }
-}
-
-
-void PolarPattern::onChanged(const App::Property* prop)
-{
-    if (prop == &Mode) {
-        auto mode = static_cast<PolarPatternMode>(Mode.getValue());
-        setReadWriteStatusForMode(mode);
-    }
-
-    Transformed::onChanged(prop);
-}
-
-void PolarPattern::setReadWriteStatusForMode(PolarPatternMode mode)
-{
-    Offset.setReadOnly(mode != PolarPatternMode::Spacing);
-    Angle.setReadOnly(mode != PolarPatternMode::Extent);
-}
-
-
-void PolarPattern::updateSpacings()
-{
-    std::vector<double> spacings = Spacings.getValues();
-    size_t targetCount = Occurrences.getValue() - 1;
-
-    for (auto& spacing : spacings) {
-        if (spacing == Offset.getValue()) {
-            spacing = -1.0;
-        }
-    }
-
-    if (spacings.size() == targetCount) {
-        return;
-    }
-    if (spacings.size() < targetCount) {
-        spacings.reserve(targetCount);
-        while (spacings.size() < targetCount) {
-            spacings.push_back(-1.0);
-        }
-    }
-    else {
-        spacings.resize(targetCount);
-    }
-
-    Spacings.setValues(spacings);
 }
 
 }  // namespace PartDesign
