+++ src/Mod/PartDesign/Gui/TaskPatternParameters.cpp
@@ -172,7 +172,7 @@ void TaskPatternParameters::bindProperties()
         parametersWidget->bindProperties(
             &linear->Direction,
             &linear->Reversed,
-            &linear->Mode,
+            &linear->Part::LinearPatternExtension::Mode,
             &linear->Length,
             &linear->Offset,
             &linear->Spacings,
@@ -197,7 +197,7 @@ void TaskPatternParameters::bindProperties()
         parametersWidget->bindProperties(
             &polar->Axis,
             &polar->Reversed,
-            &polar->Mode,
+            &polar->Part::PolarPatternExtension::Mode,
             &polar->Angle,
             &polar->Offset,
             &polar->Spacings,
@@ -443,7 +443,7 @@ void TaskPatternParameters::updateSpacingLabels()
             // calculateOffsetVector returns a vector whose magnitude is one step in 'Extent' mode,
             // or a normalized vector in 'Spacing' mode. We normalize it to be safe.
             gp_Vec offset1 = linearPattern->calculateOffsetVector(
-                PartDesign::LinearPatternDirection::First
+                Part::LinearPatternDirection::First
             );
             Base::Vector3d direction1(0, 0, 1);  // Default direction
             if (offset1.Magnitude() > Precision::Confusion()) {
@@ -458,7 +458,7 @@ void TaskPatternParameters::updateSpacingLabels()
         // Handle second direction widget
         if (parametersWidget2) {
             gp_Vec offset2 = linearPattern->calculateOffsetVector(
-                PartDesign::LinearPatternDirection::Second
+                Part::LinearPatternDirection::Second
             );
             Base::Vector3d direction2(0, 1, 0);  // Default direction
             if (offset2.Magnitude() > Precision::Confusion()) {
